ai_agent/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
groq
google-generativeai
anthropic
PyYAML
tiktoken